Refacto du système d'ingestion (#3076)
