+++ server/src/common/model/@types/EffectifsQueue.ts
@@ -68,6 +68,7 @@ export interface EffectifsQueue {
   formation_cfd?: any; // Code CFD de la formation
   source: string; // Source du dossier apprenant (Ymag, Gesti, TDB_MANUEL, TDB_FILE...)
   effectif_id?: any; // Id de l'effectif associé, objectId
+  organisme_id?: any; // Id de l'organisme associé, objectId
   updated_at?: Date; // Date de mise à jour en base de données
   created_at: Date; // Date d'ajout en base de données
   processed_at?: Date; // Date de process des données
