+++ server/src/jobs/ingestion/process-ingestion.ts
@@ -1,0 +1,263 @@
+/* eslint-disable @typescript-eslint/ban-types */
+import { PromisePool } from "@supercharge/promise-pool";
+import { ObjectId } from "mongodb";
+import pPipe from "p-pipe";
+
+import {
+  insertEffectif,
+  lockEffectif,
+  updateEffectif,
+  addEffectifComputedFields,
+  mergeEffectifWithDefaults,
+} from "@/common/actions/effectifs.actions";
+import {
+  buildNewHistoriqueStatutApprenant,
+  mapEffectifQueueToEffectif,
+  completeEffectifAddress,
+  checkIfEffectifExists,
+} from "@/common/actions/engine/engine.actions";
+import { isOrganismeFiableForCouple } from "@/common/actions/engine/engine.organismes.utils";
+import {
+  findOrganismeByUaiAndSiret,
+  setOrganismeTransmissionDates,
+} from "@/common/actions/organismes/organismes.actions";
+import logger from "@/common/logger";
+import { Effectif } from "@/common/model/@types";
+import { EffectifsQueue } from "@/common/model/@types/EffectifsQueue";
+import { effectifsQueueDb } from "@/common/model/collections";
+import { formatError } from "@/common/utils/errorUtils";
+import { sleep } from "@/common/utils/timeUtils";
+import dossierApprenantSchemaV1V2 from "@/common/validation/dossierApprenantSchemaV1V2";
+
+const sleepDuration = 10_000;
+const NB_ITEMS_TO_PROCESS = 100;
+type Options = { id?: string; force?: boolean };
+
+/**
+ * Fonction de process de la file d'attente des effectifs en boucle
+ * @param options
+ */
+export const processEffectifsQueueEndlessly = async (options: Options) => {
+  logger.info(`Process effectifs queue with options ${JSON.stringify(options)}`);
+
+  // the inc is here to avoid infinite loop and memory leak
+  let inc = 0;
+  do {
+    const { totalProcessed } = await processEffectifsQueue(options);
+    if (sleepDuration && totalProcessed === 0 && !options.id) {
+      await sleep(sleepDuration);
+    }
+    inc++;
+  } while (!options.id && inc < 100);
+};
+
+/**
+ * Fonction de process unitaire de la file d'attente des effectifs
+ * @param options
+ * @returns
+ */
+export const processEffectifsQueue = async (options?: Options) => {
+  const { id, force } = options || { id: undefined, force: false };
+
+  const filter: Record<string, any> = force ? {} : { processed_at: { $exists: false } };
+  if (id) {
+    filter._id = id;
+  }
+
+  const result = await processItems(filter);
+  return result;
+};
+
+type ProcessItemsResult = {
+  totalProcessed: number;
+  totalValidItems: number;
+  totalInvalidItems: number;
+};
+
+/**
+ * Fonction de process des 100 derniers éléments de la queue
+ * @param filter
+ */
+async function processItems(filter: any): Promise<ProcessItemsResult> {
+  const total = await effectifsQueueDb().countDocuments(filter);
+  let totalValidItems = 0;
+
+  const itemsToProcess = await effectifsQueueDb()
+    .find(filter)
+    .sort({ created_at: 1 })
+    .limit(NB_ITEMS_TO_PROCESS)
+    .toArray();
+
+  logger.info(
+    { filter, count: itemsToProcess.length, total },
+    `${itemsToProcess.length}/${total} items à traiter (avec filtre ${JSON.stringify(filter)}})`
+  );
+
+  await PromisePool.withConcurrency(10)
+    .for(itemsToProcess)
+    .process(async (effectifQueued, index) => {
+      try {
+        const startDate = new Date();
+        let effectifQueueToUpdate: Partial<EffectifsQueue> = {};
+
+        // Phase de contrôle de l'organisme
+        const isOrganismeFiable = await controlOrganismeFiable(
+          effectifQueued.uai_etablissement,
+          effectifQueued.siret_etablissement,
+          effectifQueueToUpdate
+        );
+
+        if (isOrganismeFiable) {
+          // Phase de transformation d'une donnée de queue
+          const { effectif } = await transformEffectifQueueItem(effectifQueued, effectifQueueToUpdate);
+
+          if (effectif) {
+            totalValidItems++;
+
+            // Phase d'ajout ou update d'un effectif
+            await createOrUpdateEffectif(effectif, effectifQueueToUpdate);
+          }
+        }
+
+        // MAJ de la queue pour indiquer que les données ont été traitées
+        await effectifsQueueDb().updateOne(
+          { _id: effectifQueued._id },
+          { $set: { ...effectifQueueToUpdate, processed_at: new Date() } }
+        );
+
+        const durationInMs = new Date().getTime() - startDate.getTime();
+        logger.info(
+          `#${index} Item ${
+            effectifQueued._id
+          } created at ${effectifQueued.created_at?.toISOString()} processed in ${durationInMs} ms`,
+          { duration: durationInMs }
+        );
+
+        return {
+          validItem: effectifQueueToUpdate.validation_errors?.length === 0 && !effectifQueueToUpdate.error,
+          error: undefined,
+        };
+      } catch (err: any) {
+        logger.error({ err, index }, `an error occured while processing effectif queue item ${index}: ${err.message}`);
+        logger.error(err);
+        return { validItem: false, error: true };
+      }
+    });
+
+  logger.info(`${itemsToProcess.length} items traités`);
+  if (itemsToProcess.length > 0) {
+    logger.info(`${totalValidItems} items valides`);
+    logger.info(`${itemsToProcess.length - totalValidItems} items invalides`);
+  }
+
+  return {
+    totalProcessed: itemsToProcess.length,
+    totalValidItems,
+    totalInvalidItems: itemsToProcess.length - totalValidItems,
+  };
+}
+
+/**
+ * Fonction contrôle de la fiabilité de l'organisme via le couple UAI / SIRET reçu
+ * @param uai
+ * @param siret
+ * @returns
+ */
+const controlOrganismeFiable = async (
+  uai: string | undefined,
+  siret: string | undefined,
+  effectifQueueToUpdate: Partial<EffectifsQueue>
+) => {
+  if (!(await isOrganismeFiableForCouple(uai, siret))) {
+    effectifQueueToUpdate.error = `Organisme (uai : ${uai} et siret : ${siret}) non fiable`;
+    return false;
+  }
+
+  return true;
+};
+
+/**
+ *
+ * @param effectifQueued
+ * @param dataToUpdate
+ * @returns
+ */
+const transformEffectifQueueItem = async (
+  effectifQueued: EffectifsQueue,
+  effectifQueueToUpdate: Partial<EffectifsQueue>
+) => {
+  let result;
+  let effectif;
+
+  try {
+    // Vérification schéma & transformation en 2 objets effectif & organisme
+    result = await dossierApprenantSchemaV1V2()
+      .transform(async (data) => ({
+        effectif: await pPipe(mapEffectifQueueToEffectif, mergeEffectifWithDefaults, completeEffectifAddress)(data),
+        organisme: await pPipe(
+          () => findOrganismeByUaiAndSiret(effectifQueued?.uai_etablissement, effectifQueued?.siret_etablissement),
+          setOrganismeTransmissionDates
+        )(data),
+      }))
+      .safeParseAsync(effectifQueued);
+
+    if (result.error) {
+      effectifQueueToUpdate.validation_errors =
+        result.error?.issues.map(({ path, message }) => ({ message, path })) || [];
+    } else {
+      const { effectif: effectifData, organisme } = result.data as any;
+
+      // Complétion de l'objet effectif avec l'organisme id et le computed info
+      effectif = {
+        ...effectifData,
+        organisme_id: organisme._id,
+        _computed: addEffectifComputedFields(organisme),
+      };
+
+      // Complétion de l'objet dans la queue avec l'organisme id
+      effectifQueueToUpdate.organisme_id = organisme._id;
+    }
+  } catch (err) {
+    effectifQueueToUpdate.error = formatError(err).message;
+  }
+
+  return { effectif };
+};
+
+/**
+ * Fonction de création ou de MAJ de l'effectif depuis la queue
+ * @param effectif
+ * @param dataToUpdate
+ */
+const createOrUpdateEffectif = async (effectif: Effectif, dataToUpdate: Partial<EffectifsQueue>) => {
+  try {
+    let effectifId: ObjectId;
+    const found = await checkIfEffectifExists(effectif);
+
+    // Gestion des MAJ d'effectif
+    if (found) {
+      effectifId = found._id;
+
+      // Update de l'historique
+      effectif.apprenant.historique_statut = buildNewHistoriqueStatutApprenant(
+        found.apprenant.historique_statut,
+        effectif.apprenant?.historique_statut[0]?.valeur_statut,
+        effectif.apprenant?.historique_statut[0]?.date_statut
+      );
+      const updatedEffectif = await updateEffectif(effectifId, effectif);
+      if (updatedEffectif) {
+        await lockEffectif(updatedEffectif);
+      }
+    } else {
+      const effectifCreated = await insertEffectif(effectif);
+      effectifId = effectifCreated._id;
+      await lockEffectif(effectifCreated);
+    }
+
+    // Complétion de l'objet dans la queue avec l'effectif id
+    dataToUpdate.effectif_id = effectifId;
+  } catch (e: any) {
+    const err = formatError(e);
+    dataToUpdate.error = err.toString();
+  }
+};
